protocol: hoist GameId + POST /games DTOs from server
GameId and the create-game request/response are client/server contract,
not server-only. Move them to the shared protocol crate so the web client
reuses the real types. Id minting (uuid) stays server-side as
id::random_game_id; server re-exports protocol::GameId so existing paths
and tests are untouched.

Refs #185.
